Update Joystick.c
Hopefully this fixes the entire image being offset one pixel to the left.  This was due to moving left before the first pixel is drawn.  I am really tired right now (was up all night) and cannot test it, but will once I get some much needed sleep.

I'm not sure what the weirdness is going on in the upper left.  It's easy enough to fix by hand, but it would be really interesting to figure out what that is.

Finally, when it gets to the second to last line, it doesn't move down and draws the last line over it.  This I am not entirely sure about either and warrants further investigation because it is not as easy to fix by hand.