Fixed the printing offset issue
It was mainly due to RX and RY sticks set to zero in the report (also with a physical joycon you get the same bounce back arounf the edges).

I had to revert the printing direction to avoid a pixel skipping issue (also this one was present with a physical joycon).

I cleaned up printing states, adding one to syncronize the starting position.

Finally, I replaced the "braking" concept with an "echo" one, cheking the time with a spin loop. This seems more robust to skipping pixels issues (IMHO mainly coming from the Switch being busy in processing something else).