Tuning on timing, and a fix on double variable declaration
With 40 ms of echo wait time you may still see some pixel skipping, expecially when printing lines with some white or black continous areas (not a big issue for dithered images, but still something worth to fix). Is a bit slower :-(.

image_data declared as external (as it should formally be), to improve compiler compatibility.